added sa separate receipt page for virtual terminal

+++ cmd/web/render.go
@@ -30,7 +30,7 @@ var functions = template.FuncMap{
 
 // formatCurrency takes widget.Price in cent to the price in dollar
 func formatCurrency(n int) string {
-	f := float32(n / 100)
+	f := float32(n) / float32(100)
 
 	return fmt.Sprintf("$%.2f", f)
 }
